sar_oil_slick_detection/requirements.txt
streamlit>=1.28.0
torch>=2.0.0
torchvision>=0.15.0
huggingface-hub>=0.17.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
matplotlib>=3.7.0
